pyproject.toml
[project]
name = "ai-news-aggregator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "docling>=2.84.0",
    "feedparser>=6.0.12",
    "google-genai>=1.73.1",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "youtube-transcript-api>=1.2.4",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
